Remove latest-version script
This is the forward-port of the changes we did for the release-1.18
branch as well.
